Update local config and matching logic
I'm not sure why I am needing to do this. It looks like the
rust-lang/rustfmt README.md file was force pushed to an older
revision. For instance, the minimum Rust version was updated to
1.31 in rust-lang/rustfmt#4554, but that commit
(rust-lang/rustfmt@769fe94) is no longer on the rust-lang/rustfmt
`master` branch. The latest commit on the README.md file in that
repository is now rust-lang/rustfmt@dfd276, which was committed on
Aug 15, 2019.